Move environment resolvers to plugin
